patch to fix warning in timeval (thanks Adam) (cherry picked from commit 5ead62378052b49f6657de431d4ac9304c3efdd4) Original: 361d1a297d18dab649e157010cfa4bb72ea3d33b

+++ lalapps/src/ring/coh_PTF_utils.c
@@ -1060,10 +1060,10 @@ void timeval_print(struct timeval *tv)
   char buffer[30];
   time_t curtime;
 
-  printf("%ld.%06ld", tv->tv_sec, tv->tv_usec);
+  printf("%lu.%06lu", (unsigned long)tv->tv_sec, (unsigned long)tv->tv_usec);
   curtime = tv->tv_sec;
   strftime(buffer, 30, "%m-%d-%Y  %T", localtime(&curtime));
-  printf(" = %s.%06ld\n", buffer, tv->tv_usec);
+  printf(" = %s.%06lu\n", buffer, (unsigned long)tv->tv_usec);
 }
 
 /*
